refactor: moved data class in separate files in models folder
