+++ lib/src/recipe_controller.dart
@@ -3,22 +3,23 @@ import 'package:html/parser.dart' show parse;
 import 'package:http/http.dart' as http;
 import 'package:optional/optional.dart';
 
+import 'models/recipe.dart';
+import 'models/recipe_parsing_job.dart';
 import 'recipe-scripts/kptncook.dart';
-import 'recipe_models.dart' show Recipe, RecipeInfo;
 
 final Map<String, Optional<Recipe> Function(Document, int)>
     _recipeParseMethodMap = {
   'mobile.kptncook.com': parseKptnCookRecipe,
 };
 
-/// Collects recipes from the websites in the passed [recipeInfos].
+/// Collects recipes from the websites in the passed [recipeParsingJobs].
 ///
-/// For each [RecipeInfo] a http request is made to the website containing the
-/// recipe. From there the recipe is parsed and adjusted to the amount of
+/// For each [RecipeParsingJob] a http request is made to the website containing
+/// the recipe. From there the recipe is parsed and adjusted to the amount of
 /// servings. The list of the parsed [Recipe]s is returned.
 /// [language] is set as 'Accept-Language' header in each http request.
 Future<List<Recipe>> collectRecipes(
-  List<RecipeInfo> recipeInfos,
+  List<RecipeParsingJob> recipeParsingJobs,
   String language,
 ) async {
   var results = <Recipe>[];
@@ -28,8 +29,8 @@ Future<List<Recipe>> collectRecipes(
     'Accept-Language': language,
   };
 
-  for (var recipe in recipeInfos) {
-    await _collectRecipe(client, recipe, headers).then(
+  for (var recipeParsingJob in recipeParsingJobs) {
+    await _collectRecipe(client, recipeParsingJob, headers).then(
       (optionalRecipe) => optionalRecipe.ifPresent(results.add),
     );
   }
@@ -40,16 +41,16 @@ Future<List<Recipe>> collectRecipes(
 
 Future<Optional<Recipe>> _collectRecipe(
   http.Client client,
-  RecipeInfo recipe,
+  RecipeParsingJob recipeParsingJob,
   Map<String, String> headers,
 ) async {
-  var response = await client.get(recipe.url, headers: headers);
+  var response = await client.get(recipeParsingJob.url, headers: headers);
 
   var document = parse(response.body);
 
-  var parseMethod = _recipeParseMethodMap[recipe.url.host];
+  var parseMethod = _recipeParseMethodMap[recipeParsingJob.url.host];
 
-  return parseMethod!.call(document, recipe.servings);
+  return parseMethod!.call(document, recipeParsingJob.servings);
 }
 
 /// Checks if the passed [Uri] is supported.
